Split link page into ShortLinkPage and LinkAdvancedPage with routing
