+++ src/App.tsx
@@ -24,7 +24,8 @@ import MarketplaceModule from './pages/MarketPlace/MarketplaceModule';
 import MarketplacePostDetail from './pages/MarketPlace/MarketplacePostDetail';
 import MarketplaceStatistics from './pages/MarketPlace/MarketplaceStatistics';
 import SettingPage from './pages/Setting/SettingPage';
-import LinkPage from './pages/Link/LinkPage';
+import ShortLinkPage from './pages/Link/ShortLinkPage';
+import LinkAdvancedPage from './pages/Link/LinkAdvancedPage';
 
 // Component that renders routes based on user role
 const AppRoutes = () => {
@@ -209,7 +210,13 @@ const AppRoutes = () => {
 
       <Route path="/link" element={
         <ProtectedRoute>
-          <LinkPage />
+          <ShortLinkPage />
+        </ProtectedRoute>
+      } />
+
+      <Route path="/link-advanced" element={
+        <ProtectedRoute>
+          <LinkAdvancedPage />
         </ProtectedRoute>
       } />
