+++ src/pages/Link/index.js
@@ -1,3 +1,5 @@
-export { default as LinkPage } from './LinkPage';
+export { default as ShortLinkPage } from './ShortLinkPage';
+export { default as LinkAdvancedPage } from './LinkAdvancedPage';
+export { default as LinkPage } from './LinkPage'; // Keep for backward compatibility
 export { default as LinkAnalytics } from './LinkAnalytics';
 export { default as AnalyticsSummary } from './AnalyticsSummary';
